fix(ios): update OneSignal SDK to 2.9.3
