+++ ios/Cartfile.resolved
@@ -1,0 +1,1 @@
+github "OneSignal/OneSignal-iOS-SDK" "2.9.3"
