+++ ios/module.xcconfig
@@ -25,6 +25,4 @@
 //
 // IMPORTANT NOTE: always use $(inherited) in your overrides
 //
-FRAMEWORK_SEARCH_PATHS=$(SRCROOT)/../../modules/iphone/com.williamrijksen.onesignal/2.0.1/platform "~/Library/Application\ Support/Titanium/modules/iphone/com.williamrijksen.onesignal/2.0.1/platform"
-OTHER_LDFLAGS=$(inherited) -framework OneSignal -framework UserNotifications
-LD_RUNPATH_SEARCH_PATHS= $(inherited) "@executable_path/Frameworks" $(FRAMEWORK_SEARCH_PATHS)
+OTHER_LDFLAGS=$(inherited) -framework UserNotifications
